Adjust return types in UserData* to deal with conversion warnings

+++ source/shared/loading/iuserdata.h
@@ -31,8 +31,8 @@ public:
 
     virtual QString firstVectorName() const = 0;
 
-    virtual int numUserDataVectors() const = 0;
-    virtual int numValues() const = 0;
+    virtual size_t numUserDataVectors() const = 0;
+    virtual size_t numValues() const = 0;
 
     virtual const std::vector<QString>& vectorNames() const = 0;
