+++ source/shared/loading/userdatavector.h
@@ -61,9 +61,9 @@ public:
     QStringList toStringList() const;
 
     const QString& name() const { return _name; }
-    int numValues() const { return static_cast<int>(_values.size()); }
-    int numUniqueValues() const;
-    void reserve(int size) { _values.reserve(size); }
+    size_t numValues() const { return _values.size(); }
+    size_t numUniqueValues() const;
+    void reserve(size_t size) { _values.reserve(size); }
 
     int intMin() const { return _intMin; }
     int intMax() const { return _intMax; }
